relocate status-bar code
