+++ lib/services/status-bar/signal-list-view.js
@@ -1,46 +1,9 @@
 /* @flow */
 
-import React from "react";
-import { Disposable } from "atom";
 import SelectListView from "atom-select-list";
 
-import WSKernel from "../ws-kernel";
-import kernelManager from "../kernel-manager";
-import StatusBar from "../components/status-bar";
-
-import { log, reactFactory } from "../utils";
-
-import type { Store } from "../store";
-import type Kernel from "../kernel";
-
-export function addStatusBar(
-  store: Store,
-  statusBar: atom$StatusBar,
-  handleKernelCommand: Function
-) {
-  const statusBarElement = document.createElement("div");
-  statusBarElement.classList.add("inline-block", "hydrogen");
-
-  const statusBarTile = statusBar.addLeftTile({
-    item: statusBarElement,
-    priority: 100
-  });
-
-  const onClick = showKernelCommands.bind(this, handleKernelCommand);
-
-  reactFactory(<StatusBar store={store} onClick={onClick} />, statusBarElement);
-
-  return new Disposable(() => statusBarTile.destroy());
-}
-
-function showKernelCommands(handleKernelCommand, kernel) {
-  if (!this.signalListView) {
-    this.signalListView = new SignalListView(kernel);
-    this.signalListView.onConfirmed = (kernelCommand: { command: string }) =>
-      handleKernelCommand(kernelCommand);
-  }
-  this.signalListView.toggle();
-}
+import WSKernel from "../../ws-kernel";
+import { log } from "../../utils";
 
 const basicCommands = [
   { name: "Interrupt", value: "interrupt-kernel" },
@@ -53,7 +16,7 @@ const wsKernelCommands = [
   { name: "Disconnect from", value: "disconnect-kernel" }
 ];
 
-class SignalListView {
+export default class SignalListView {
   onConfirmed: ?(command: { command: string }) => void;
   panel: ?atom$Panel;
   previouslyFocusedElement: ?HTMLElement;
