+++ lib/services/status-bar/status-bar.js
@@ -3,8 +3,8 @@
 import React from "react";
 import { observer } from "mobx-react";
 
-import type Kernel from "../kernel";
-import type { Store } from "../store";
+import type Kernel from "../../kernel";
+import type { Store } from "../../store";
 
 type Props = {
   store: Store,
